requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
Werkzeug==3.0.1
PyJWT==2.8.0
bcrypt==4.1.1
requests==2.31.0
python-dateutil==2.8.2
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.7
phonenumbers==8.13.26
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
boto3==1.34.10
azure-storage-blob==12.27.1
stripe==7.9.0
python-slugify==8.0.1
email-validator==2.1.0

